requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via django
django-crispy-forms==1.9.2  # via -r requirements.in
django-extensions==3.0.9  # via -r requirements.in
django-querycount==0.7.0  # via -r requirements.in
django==3.1.1             # via -r requirements.in
python-nmap==0.6.1        # via -r requirements.in
pytz==2020.1              # via django
sqlparse==0.3.1           # via django
